submissions/Spark Vest Investment/agent-server/requirements.txt
agno==1.1.1
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
curl-cffi==0.11.1
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.16
duckduckgo-search==7.3.2
email-validator==2.2.0
exa-py==1.13.1
fastapi==0.115.8
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
finnhub-python==2.4.22
flask==3.1.0
frozendict==2.4.6
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.38.0
google-genai==1.5.0
groq==0.18.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.10.0
lxml==5.3.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
numpy==2.2.3
openai==1.79.0
packaging==25.0
pandas==2.2.3
peewee==3.17.9
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.6.0
primp==0.12.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
pygments==2.19.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
time-machine==2.16.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
webencodings==0.5.1
websockets==14.2
werkzeug==3.1.3
yfinance==0.2.61
